requirements.txt
tyro
wandb
flax
tqdm
numpy
tiktoken
datasets
tensorflow
transformers
optax
einops
psgd-jax
chex